User/requirements.txt
streamlit
langchain
langchain_aws
qdrant-client
boto3
python-dotenv